requirements.txt
annotated-types==0.7.0
anyio==4.4.0
audioread==3.0.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
mutagen==1.47.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.45.0
openai-whisper==20231117
packaging==24.1
platformdirs==4.3.3
pooch==1.8.2
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.2
threadpoolctl==3.5.0
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.5
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.0.1
youtube-transcript-api==0.6.2
yt-dlp==2024.8.6
